[quant][fx][graphmode] Fix a bug for sub
Summary:
Previously for sub we set the dtype for sub with qconfig since it's matched with a QuantizeHandler,
however this is incorrect, the dtype for sub is decided by whether the output is quantized or not,
so we added a check of is_output_quantized while deciding the dtype for the output of sub.

Later: is_output_quantized now depends on is_reference, which is pretty confusing and it may cause problems down the road, we should remove this dependency in the future.

Test Plan:
python test/test_quantization.py TestQuantizeFx.test_sub_scalar

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: d93100633bd943ae65f38bb5b289d7747f7783a0
Pull Request resolved: https://github.com/pytorch/pytorch/pull/65109